Qua: add SV normalization/denormalization
Adds SV normalization and denormalization routines to Qua.

Additionally, a new .qua field is added, BPMDoesNotAffectScrollVelocity.
When this is set to true, SVs are assumed to be normalized in the .qua
itself. If this flag is set before serializing, the .qua will be saved
with normalized SVs.

The normalization and denormalization routines were converted from the
plitki [1] implementation, which was very well tested. All tests were
converted from there too.

[1]: https://github.com/YaLTeR/plitki/blob/41e9dd034848222d5047524bd1a5fe74e1f88dab/plitki-map-qua/src/lib.rs#L289