Implement continuity adjustment
Having a section where notes are relatively easy, compared to the hardest sections of the map, drops the rating way more than it should.
This part computes a "continuity" value, the higher it is the less "easy" notes it has.
Maps with a higher continuity should get nerfed, since they are overrated right now.

The continuity is computed by taking a comparison value, set as the average of the hardest 40% of all sections.
All sections are compared to that value, the average is the continuity.
A square root is used to decrease the gap between easy sections and hard sections and considered the core of this adjustment.