Qua: Fix SV normalization with 2+ timing points and SV
This fixes a bug that I initially missed which results in different
behavior from old Quaver code and a mismatch between normalization and
denormalization when the following conditions are met:

1. BPMDoesNotAffectSV = false (old maps),
2. There's at least two timing points and at least one SV all on the
   same StartTime.

In this case the SV was already considered when converting the first
timing point, even though it only applies to the last timing point.

We ran a check across ranked maps and found none that have this
configuration. I've manually tested a number of SV maps (both Quaver and
osu! converts) and found no change in behavior.